% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_boundary.R
\name{get_city_boundary}
\alias{get_city_boundary}
\title{Get city boundary from GADM}
\usage{
get_city_boundary(city)
}
\arguments{
\item{city}{Character. City name (e.g. "Aachen"). As far as now only for German cities.}
}
\value{
sf object of the city boundary
}
\description{
Get city boundary from GADM
}
