% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hotcoolspots.R
\name{get_hotcoolspots}
\alias{get_hotcoolspots}
\title{Calculates and Plots the hot and cool spots}
\usage{
get_hotcoolspots(lst_result, hot_quantile = 0.9, cold_quantile = 0.1)
}
\arguments{
\item{lst_result}{List - Result of the function get_LST() containing the LST, the boundary, the city name and the acquisition date}

\item{hot_quantile}{Numeric - Upper quantile treshold to classify hotspots (by default 0.90 = the hottest 10\% are being considered)}

\item{cold_quantile}{Numeric - Lower quantile treshold to classify coldspots (by default = 0.10 = the coldest 10\% are being considered)}
}
\value{
Plot of Hot and Cool Spots in the given city at the acquisition date
}
\description{
Calculates the hottest and coldest pixels of the city and plots them in a ready-to-use map so areas to cool off and  areas where action is needed can be identified
}
\examples{
\dontrun{
get_hotcoolspots(
lst_result= lst_result, 
hot_quantile= 0.95,
cold_quantile= 0.05)}

}
