% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LULC.R
\name{get_LULC}
\alias{get_LULC}
\title{Gets the Land Use and Land Cover Classification}
\usage{
get_LULC(path_to_CORINE, city)
}
\arguments{
\item{path_to_CORINE}{\itemize{
\item Character: Path to the CORINE Land Cover raster (.tif)
}}

\item{city}{\itemize{
\item Character: Name of the city that is being analyzed
}}
}
\value{
List with four elements: The masked raster, the look table, the city name and the city boundary
}
\description{
Loads the CORINE Land Cover raster and masks it to the boundaries of the specified city. A lookup table is created to extracr the 44 CORINE land cover classes from the dataset.
}
\examples{
#Create the list 
lulc_result <- get_LULC(
 path_to_CORINE = "C:/Users/LaraO/EAGLE_Master/1_Semester/New R-Package/U2018_CLC2018_V2020_20u1.tif",
 city = "Aachen"

#Access the masked raster
lulc_result$clc_masked

#See which land cover classes are present within the city boundary 
lulc_result$clc_lookup_present

}
