% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LST_plotted.R
\name{LST_plotted}
\alias{LST_plotted}
\title{Plots the Land Surface Temperature}
\usage{
LST_plotted(lst_result)
}
\arguments{
\item{lst_result}{List - Result of the function get_LST() containing the LST, the boundary, the city name and the acquisition date}
}
\value{
Plot of the Land Surface Temperature
}
\description{
Creates a ready-to-use map of the Land Surface Temperature at the given date, based on the output of \code{get_LST()}
}
\examples{
\dontrun{
LST_plotted (lst_result)}

}
