% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LULC_plotted.R
\name{LULC_plotted}
\alias{LULC_plotted}
\title{Plots the Land Use / Land Cover of the city based on the CORINE Dataset}
\usage{
LULC_plotted(clc_path, city)
}
\arguments{
\item{clc_path}{Character: Path to the .tif file containing the CORINE LULC dataset}

\item{city}{Character: Name of the city that is being analyzed}
}
\value{
Plot of the LULC of the city based on ggplot2
}
\description{
Plots the Land Use / Land Cover of the city based on the CORINE Dataset
}
\examples{
\dontrun{
LULC_plotted( 
clc_path = "C:/Users/LaraO/EAGLE_Master/1_Semester/New R-Package/U2018_CLC2018_V2020_20u1.tif",
city = "Aachen")
}
}
