% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_facilities_heat.R
\name{map_facilities_heat}
\alias{map_facilities_heat}
\title{Map vulnerable facilities within LST hotspots}
\usage{
map_facilities_heat(df)
}
\description{
Creates an interactive Leaflet map showing vulnerable facilities within perviously identified LST hotspots
}
\examples{
df <- table_facilities_heat(lst_result, 0.9)
map_facilities_heat(df)

}
