% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_LULC_LST.R
\name{statistics_LULC_LST}
\alias{statistics_LULC_LST}
\title{Plots the LST boxplots per LULC Class}
\usage{
statistics_LULC_LST(lulc_data, lst_result)
}
\arguments{
\item{lulc_data}{List: Result of the function get_LULC()}

\item{lst_result}{List: Result of the function get_LST()}
}
\value{
Boxplots of the temperature of the LULC classes (ggplot2 -object)
}
\description{
Creates a plot of the temperature distribution of the LULC Classes present in the study area using boxplots
}
\examples{
\dontrun{
statistics_LULC_LST(lulc_data, lst_result)
}
}
