% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_facilities_heat.R
\name{table_facilities_heat}
\alias{table_facilities_heat}
\title{Table of vulnerable facilities in LST hotspots}
\usage{
table_facilities_heat(lst_result, percentile = 0.9)
}
\arguments{
\item{lst_result}{List. Result of \code{get_LST()}}

\item{percentile}{Numeric. Hotspot threshold (e.g. 0.9 = the upper 10\% are being considered)}
}
\value{
A data frame with facility category and name.
}
\description{
Retrieves vulnerable facilities within LST hotspots for a city
and returns a table with facility name and type. Facilities are retrieved
automatically via OSM. OSM data may be incomplete for some facility types.
}
\examples{
\dontrun{
table_facilities_heat(
  lst_result = result,
  percentile = 0.10
)
}

}
