% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_boundary.R
\name{get_city_boundary}
\alias{get_city_boundary}
\title{Gets a cities boundary}
\usage{
get_city_boundary(city)
}
\arguments{
\item{city}{Character. City name (e.g. "Aachen"). As far as now only for German cities.}
}
\value{
sf object of the city boundary
}
\description{
Retrieves the cities boundary as a sf polygon using GADM
}
\examples{
get_city_boundary("Aachen")

}
